Tried to implement mouse following. Failure.
The model sort of follows the mouse. The issue is the multitude of mesh colliders.
This makes it so that I have to create a parent gameobject instead of using the original, and warps the rotation parameters.
Additional calculations could be implemented but I've concluded they're unnecessary.
Normally I was using actual calculations for mouse drag but I decided the LookAt method would end up with a more pleasing result if properly implemented.

#21