Implemented switching between 3 different forms.
The POST route has been removed even though it exists, since its functionality will not be in use.
I should probably remove some other routes, too (for example the root route and other useless getters), but it's fine like this for now because the UI looks nicer this way.
I should include a textinput field that's readonly that shows the STP command, though.

Need to add visual display for logs as well.

This was a much bigger pain than I thought it'd be, but it's now relatively simple to create a new form and add it to the list of exceptions.

The form switching could probably be done in a separate script using data structures instead of hard-coded data, but at this point I'm so sick of Unity's UI mechanisms I just want to be done already.

#21